refactor: security feature abstractions
